pyproject.toml
[tool.poetry]
name = "podcast-translation-bot"
version = "0.1.0"
description = ""
authors = ["vladislavean <vladrollss228@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
moviepy = "^1.0.3"
pytube = "^15.0.0"
python-dotenv = "^1.0.1"
speechrecognition = "^3.10.3"
werkzeug = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
